+++ src/main/java/com/db_lab/db_lab6/controller/MatchResultsController.java
@@ -1,6 +1,7 @@
 package com.db_lab.db_lab6.controller;
 
 import com.db_lab.db_lab6.domain.MatchResults;
+import com.db_lab.db_lab6.domain.dto.MatchResultsDTO;
 import com.db_lab.db_lab6.exception.MatchResultNotFoundException;
 import com.db_lab.db_lab6.security.service.SecurityService;
 import com.db_lab.db_lab6.service.MatchResultsService;
@@ -33,7 +34,7 @@ public class MatchResultsController {
         this.securityService = securityService;
     }
 
-    @Operation(summary = "get all MatchResults(for admins)")
+    @Operation(summary = "get all MatchResults(for all users)")
     @GetMapping
     public ResponseEntity<List<MatchResults>> getMatchResults() {
         List<MatchResults> matchResults = matchResultsService.getMatchResults();
@@ -56,17 +57,28 @@ public class MatchResultsController {
     }
 
     //maybe should be deleted(or only for admins)
-    @Operation(summary = "create MatchResults (for authorized users)")
+    @Operation(summary = "create MatchResults (for admins)")
     @PostMapping
-    public ResponseEntity<HttpStatus> createMatchResults(@RequestBody MatchResults matchResults) {
-        matchResultsService.createMatchResults(matchResults);
-        return new ResponseEntity<>(HttpStatus.CREATED);
+    public ResponseEntity<HttpStatus> createMatchResults(@RequestBody MatchResultsDTO matchResultsDTO, Principal principal) {
+        if(principal == null){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+        if (securityService.checkIfAdmin(principal.getName())) {
+            matchResultsService.createMatchResults(matchResultsDTO);
+            return new ResponseEntity<>(HttpStatus.CREATED);
+        }else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
     }
 
 
     @Operation(summary = "update MatchResults (for authorized users)")
     @PutMapping
     public ResponseEntity<HttpStatus> updateMatchResults(@RequestBody MatchResults matchResults, Principal principal) {
+        if(principal == null){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         if (securityService.checkIfAdmin(principal.getName())) {
             matchResultsService.updateMatchResults(matchResults);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -78,6 +90,9 @@ public class MatchResultsController {
     @Operation(summary = "delete MatchResults (for authorized users)")
     @DeleteMapping("/{id}")
     public ResponseEntity<HttpStatus> deleteMatchResults(@PathVariable Long id, Principal principal) {
+        if(principal == null){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         if (securityService.checkIfAdmin(principal.getName())) {
             matchResultsService.deleteMatchResultsById(id);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
